make sure the .tmx file exists

+++ build_maps.sh
@@ -4,6 +4,11 @@ if [ -z "$1" ]
     echo "Missing map TMX file"
     exit 1
 fi
+
+if [ ! -f "$1" ]; then
+  echo "File $1 does not exist"
+  exit 1
+fi
 export DOCKER_DEFAULT_PLATFORM=linux/amd64
 export PROGRESS_NO_TRUNC=1
 docker kill $(docker ps -q) 2> /dev/null
